Move MarlinQQQTensor out of AQT
